many changes, can't remember everything, but these are the most important:  - attaching entities to other entities works  - displaying models, lights and shadows works  - controlling a spectator works  - removed an update hack in PositionableEntity because I've found a much better solution and with "works" I mean: it works on client, server and standalone

+++ code/branches/objecthierarchy/src/network/GamestateClient.h
@@ -47,7 +47,7 @@
 #include "packet/Gamestate.h"
 #include "GamestateHandler.h"
 
-const unsigned int GAMESTATEID_INITIAL=-1;
+const unsigned int GAMESTATEID_INITIAL = (unsigned int)-1;
 
 namespace network
 {
